agent/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "claudriel-agent"
version = "0.0.0"
description = "Claudriel chat subprocess (Anthropic tool loop + internal API tools)"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "pytest",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["claudriel_agent*"]

[tool.black]
target-version = ["py311"]
line-length = 100
extend-exclude = "/(\\.venv|\\.pytest_cache)/"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["claudriel_agent", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["claudriel_agent"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["claudriel_agent"]
warn_unused_ignores = true
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = "anthropic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

# Thin proxy tools: uniform typing deferred (TOOL_DEF + execute pattern).
[[tool.mypy.overrides]]
module = "claudriel_agent.tools.*"
ignore_errors = true

# Eval / benchmark CLIs: Anthropic response unions and loose dicts.
[[tool.mypy.overrides]]
module = "claudriel_agent.eval_benchmark"
ignore_errors = true

[[tool.mypy.overrides]]
module = "claudriel_agent.eval_contracts"
ignore_errors = true

[[tool.mypy.overrides]]
module = "claudriel_agent.eval_judge"
ignore_errors = true

[[tool.mypy.overrides]]
module = "claudriel_agent.eval_report"
ignore_errors = true

[[tool.mypy.overrides]]
module = "claudriel_agent.eval_runner"
ignore_errors = true

[[tool.mypy.overrides]]
module = "claudriel_agent.eval_schema"
ignore_errors = true
